requirements.txt
requests
regex
scikit-learn==0.22.2.post1
seaborn
rouge-score
elasticsearch
justext==2.2.0
warcio==1.7.3
nltk==3.4.5
numpy==1.18.1
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tqdm==4.42.1
